+++ VR_Tools.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35327.3
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VR_Tools_UI", "VR_Tools_UI\VR_Tools_UI.csproj", "{373B60C4-6BF9-4DF3-A4F2-51CDC268CAF7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VR_Tools", "VR_Tools\VR_Tools.csproj", "{373B60C4-6BF9-4DF3-A4F2-51CDC268CAF7}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VR_Tools_UI.Desktop", "VR_Tools_UI.Desktop\VR_Tools_UI.Desktop.csproj", "{C3770B12-DBAC-463A-ADBF-E195480817CA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VR_Tools.Desktop", "VR_Tools.Desktop\VR_Tools.Desktop.csproj", "{C3770B12-DBAC-463A-ADBF-E195480817CA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
