+++ VR_Tools/ViewModels/ViewModelBase.cs
@@ -1,6 +1,6 @@
 ﻿using ReactiveUI;
 
-namespace VR_Tools_UI.ViewModels;
+namespace VR_Tools.ViewModels;
 
 public class ViewModelBase : ReactiveObject
 {
