+++ VR_Tools/Views/MainWindow.axaml
@@ -1,10 +1,10 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:vm="using:VR_Tools_UI.ViewModels"
+        xmlns:vm="using:VR_Tools.ViewModels"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:views="clr-namespace:VR_Tools_UI.Views"
-        x:Class="VR_Tools_UI.Views.MainWindow"
+        xmlns:views="clr-namespace:VR_Tools.Views"
+        x:Class="VR_Tools.Views.MainWindow"
 		Icon="\Assets\Icon.ico"
         Title="VR Tools"
 		Width="750" Height="550"
